share AnalysisResultsController between playground and embeds
